tests(jobmanager): add job manager tests
* move actor tests from JobManagerSpec into JobManagerActorSpecs
* add job manager tests in JobManagerSpec
