+++ job-server/src/test/scala/spark/jobserver/JobWithNamedRddsSpec.scala
@@ -10,7 +10,7 @@ import org.apache.spark.rdd.RDD
 import com.typesafe.config.Config
 import spark.jobserver.common.akka.AkkaTestUtils
 
-class JobWithNamedRddsSpec extends JobSpecBase(JobManagerSpec.getNewSystem) {
+class JobWithNamedRddsSpec extends JobSpecBase(JobManagerActorSpec.getNewSystem) {
 
   val sc = new SparkContext("local[4]", getClass.getSimpleName, new SparkConf)
